+++ lib/client/weblogin.go
@@ -33,11 +33,11 @@ import (
 	"github.com/gravitational/roundtrip"
 	"github.com/gravitational/teleport"
 	"github.com/gravitational/teleport/api/client/proto"
+	"github.com/gravitational/teleport/api/client/webclient"
 	"github.com/gravitational/teleport/api/constants"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/defaults"
-	"github.com/gravitational/teleport/lib/web/ui"
 
 	"github.com/gravitational/trace"
 	"github.com/sirupsen/logrus"
@@ -442,7 +442,7 @@ func HostCredentials(ctx context.Context, proxyAddr string, insecure bool, req t
 }
 
 // GetWebConfig is used by teleterm to fetch webconfig.js from proxies
-func GetWebConfig(ctx context.Context, proxyAddr string, insecure bool) (*ui.WebConfig, error) {
+func GetWebConfig(ctx context.Context, proxyAddr string, insecure bool) (*webclient.WebConfig, error) {
 	clt, _, err := initClient(proxyAddr, insecure, nil)
 	if err != nil {
 		return nil, trace.Wrap(err)
@@ -461,12 +461,10 @@ func GetWebConfig(ctx context.Context, proxyAddr string, insecure bool) (*ui.Web
 	// WebConfig is served as JS file where GRV_CONFIG is a global object name
 	text := bytes.TrimSuffix(bytes.Replace(body, []byte("var GRV_CONFIG = "), []byte(""), 1), []byte(";"))
 
-	cfg := ui.WebConfig{}
+	cfg := webclient.WebConfig{}
 	if err := json.Unmarshal(text, &cfg); err != nil {
 		return nil, trace.Wrap(err)
 	}
 
 	return &cfg, nil
 }
-
-type WebConfig = ui.WebConfig
