+++ lib/teleterm/clusters/cluster_auth.go
@@ -20,18 +20,18 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/gravitational/teleport/api/client/webclient"
 	"github.com/gravitational/teleport/api/constants"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/client"
 	dbprofile "github.com/gravitational/teleport/lib/client/db"
 	"github.com/gravitational/teleport/lib/kube/kubeconfig"
-	web "github.com/gravitational/teleport/lib/web/ui"
 
 	"github.com/gravitational/trace"
 )
 
 // SyncAuthPreference fetches Teleport auth preferences and stores it in the cluster profile
-func (c *Cluster) SyncAuthPreference(ctx context.Context) (*web.WebConfigAuthSettings, error) {
+func (c *Cluster) SyncAuthPreference(ctx context.Context) (*webclient.WebConfigAuthSettings, error) {
 	_, err := c.clusterClient.Ping(ctx)
 	if err != nil {
 		return nil, trace.Wrap(err)
